wave2_network/packages/shared-py/pyproject.toml
[project]
name = "blindference-utils"
version = "0.1.0"
description = "Shared Blindference crypto, IPFS, and commitment helpers"
requires-python = ">=3.11"
license = { text = "BUSL-1.1" }
dependencies = [
    "pycryptodome>=3.20",
    "requests>=2.32",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["blindference_utils"]
